refactor(admin): consolidate duplicated date/CSV helpers into shared/format
Move parseUtcDate, formatDate, formatDateTime, and downloadCSV (22 copies
across 8 tab files, verified byte-identical) into the existing
admin/shared/format.ts module, replacing local definitions with imports.

Also retires three drifted variants (ApiKeysTab's fmtDate, ComplianceTab's
formatDateTime, KnowledgeBasesTab's formatDate) in favor of the shared
version, which folds in KnowledgeBasesTab's Number.isNaN guard and fixes
the negative-UTC-offset bug the two others lacked (appending "Z" to a
timestamp that already has a "-HH:MM" offset produced Invalid Date on
API-key expiry and compliance retention dates).

Visible micro-changes: null placeholder standardized on the em-dash "—"
(matching the pre-existing tabs) instead of the mixed "-"/"—" before, and
formatDateTime now includes the year.

Note: ApiKeysTab's fmtDate actually formatted a full date+time
(month/day/year/hour/minute), not a date-only value, despite its name — it
is now wired to the shared formatDateTime (not formatDate) to preserve
that behavior.

CHART_COLORS and HISTORY_PAGE_SIZE are left as module-private, single-use
constants, unchanged.
