+++ frontend/src/components/admin/DemoTab.tsx
@@ -18,34 +18,8 @@ import type { DemoAdminStats, DemoApplication as DemoApp, PostExperienceResponse
 import { POST_SURVEY_FIELDS } from '../survey/postSurveyFields'
 import { PRE_SURVEY_FIELDS } from '../../pages/Demo'
 import { SurveyFieldRenderer } from '../survey/SurveyFieldRenderer'
+import { downloadCSV, formatDate } from './shared/format'
 import { SearchInput } from './shared/primitives'
-
-function parseUtcDate(d: string): Date {
-  // Backend stores UTC but may omit timezone suffix; ensure JS treats it as UTC
-  if (!d.endsWith('Z') && !d.includes('+') && !d.includes('-', 10)) return new Date(d + 'Z')
-  return new Date(d)
-}
-
-function formatDate(d: string | null): string {
-  if (!d) return '-'
-  return parseUtcDate(d).toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
-}
-
-function downloadCSV(filename: string, headers: string[], rows: (string | number | null)[][]) {
-  const escape = (v: string | number | null) => {
-    if (v === null || v === undefined) return ''
-    const s = String(v)
-    return s.includes(',') || s.includes('"') || s.includes('\n') ? `"${s.replace(/"/g, '""')}"` : s
-  }
-  const csv = [headers.join(','), ...rows.map(r => r.map(escape).join(','))].join('\n')
-  const blob = new Blob([csv], { type: 'text/csv' })
-  const url = URL.createObjectURL(blob)
-  const a = document.createElement('a')
-  a.href = url
-  a.download = filename
-  a.click()
-  URL.revokeObjectURL(url)
-}
 
 function DemoResponseDetail({ responses }: { responses: Record<string, unknown> }) {
   if (!responses || Object.keys(responses).length === 0) {
