+++ frontend/src/components/admin/shared/format.ts
@@ -13,3 +13,39 @@ export function formatDuration(ms: number | null): string {
   const remainSecs = Math.round(secs % 60)
   return `${mins}m ${remainSecs}s`
 }
+
+export function parseUtcDate(d: string): Date {
+  // Backend stores UTC but may omit timezone suffix; ensure JS treats it as UTC
+  if (!d.endsWith('Z') && !d.includes('+') && !d.includes('-', 10)) return new Date(d + 'Z')
+  return new Date(d)
+}
+
+export function formatDate(d: string | null): string {
+  if (!d) return '—'
+  const date = parseUtcDate(d)
+  if (Number.isNaN(date.getTime())) return '—'
+  return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
+}
+
+export function formatDateTime(d: string | null): string {
+  if (!d) return '—'
+  const date = parseUtcDate(d)
+  if (Number.isNaN(date.getTime())) return '—'
+  return date.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric', hour: 'numeric', minute: '2-digit' })
+}
+
+export function downloadCSV(filename: string, headers: string[], rows: (string | number | null)[][]) {
+  const escape = (v: string | number | null) => {
+    if (v === null || v === undefined) return ''
+    const s = String(v)
+    return s.includes(',') || s.includes('"') || s.includes('\n') ? `"${s.replace(/"/g, '""')}"` : s
+  }
+  const csv = [headers.join(','), ...rows.map(r => r.map(escape).join(','))].join('\n')
+  const blob = new Blob([csv], { type: 'text/csv' })
+  const url = URL.createObjectURL(blob)
+  const a = document.createElement('a')
+  a.href = url
+  a.download = filename
+  a.click()
+  URL.revokeObjectURL(url)
+}
